# Improve performance of select() and addSelect()