--- a/framework/db/Query.php
+++ b/framework/db/Query.php
@@ -656,7 +656,7 @@ PATTERN;
         foreach ($columns as $columnAlias => $columnDefinition) {
             if (is_string($columnAlias)) {
                 $select[$columnAlias] = $columnDefinition;
-            } else if (preg_match('/^\w+$/', $columnDefinition)) {
+            } else if (is_string($columnDefinition) && preg_match('/^\w+$/', $columnDefinition)) {
                 $select[$columnDefinition] = $columnDefinition;
             } else {
                 $select[] = $columnDefinition;