We just profiled a 25.8s request with Blackfire, and found that 11.5s / 44.5% was spent on `yii\db\Query::getUniqueColumns()` (called by `select()` and `addSelect()`), and its callee, `getUnaliasedColumnsFromSelect()`.

<img width="423" alt="Screen Shot 2019-06-06 at 3 21 25 PM" src="https://user-images.githubusercontent.com/47792/59070991-b7551800-8871-11e9-866a-ba756773c31c.png">

[View on Blackfire ➡️](https://blackfire.io/profiles/26b68fc8-ccb8-424c-a9b6-06010fbb06e9/graph?settings%5Bdimension%5D=wt&settings%5Bdisplay%5D=landscape&settings%5BtabPane%5D=nodes&selected=yii%5Cdb%5CQuery%3A%3AgetUniqueColumns%3F34f7364a45ee985db52f3b3a2c668108&callname=main())

We rewrote `select()` and `addSelect()` to be more efficient, avoiding `getUniqueColumns()` altogether and ensuring uniqueness by normalizing SELECT columns similar to the way `orderBy()` / `normalizeOrderBy()` works. Now aliases are always normalized into the `['alias' => 'column']` syntax, and even un-aliased _column names_ (excluding expressions, sub-selects, and strings with parentheses) are aliased to themselves (`['column' => 'column']`).

The result:  `select()` isn’t even showing up on Blackfire’s radar anymore, and `addSelect()` is only taking up 663ms (17.3x faster), bringing the request time down to 14.1s (1.8x faster).

<img width="425" alt="Screen Shot 2019-06-06 at 3 53 03 PM" src="https://user-images.githubusercontent.com/47792/59071380-3565ee80-8873-11e9-9de8-413ef2574b3f.png">

[View on Blackfire ➡️](https://blackfire.io/profiles/62b0108f-2492-406b-8352-a6c8bfd8cc14/graph?settings%5Bdimension%5D=wt&settings%5Bdisplay%5D=landscape&settings%5BtabPane%5D=nodes&selected=craft%5Cdb%5CQuery%3A%3AaddSelect&callname=main())

We hope that is a performance improvement worth considering :)

| Q             | A
| ------------- | ---
| Is bugfix?    | no
| New feature?  | no
| Breaks BC?    | no
| Tests pass?   | yes