@brandonkelly 2019-06-11
Awesome! The only change we had to make in Craft was due to an `array_splice()` call on `$query->select()`, as the key that was passed in was no longer numeric. Doubt many people are doing things like that with normal queries. Probably worth mentioning that `yii\db\Query::select()` and `addSelect()` now normalize the columns when saving them to `$this->select`, so code that works directly with that property may need to be modified.